Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
pydantic
pandas